+++ apps/starter/CHANGELOG.md
@@ -1,5 +1,17 @@
 # stoker
 
+## 0.4.240
+
+### Patch Changes
+
+- Updated dependencies
+    - @stoker-platform/types@0.5.57
+    - @stoker-platform/web-app@0.5.200
+    - @stoker-platform/cli@0.5.125
+    - @stoker-platform/node-client@0.5.78
+    - @stoker-platform/utils@0.5.69
+    - @stoker-platform/web-client@0.5.81
+
 ## 0.4.239
 
 ### Patch Changes