+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @stoker-platform/cli
 
+## 0.5.125
+
+### Patch Changes
+
+- Updated dependencies
+    - @stoker-platform/types@0.5.57
+    - @stoker-platform/node-client@0.5.78
+    - @stoker-platform/utils@0.5.69
+
 ## 0.5.124
 
 ### Patch Changes